dealing with C++ warnings, also:
API CHANGE? some of the command-line options in unrrdu have change
type to reflect underlying type changes in nrrd, but there is no
functionality change.  The "-a" option to identify an axis, for
example, has changed from "int" to "unsigned int".


git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@2729 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/unrrdu/1op.c
@@ -81,7 +81,7 @@ unrrdu_1opMain(int argc, char **argv, char *me, hestParm *hparm) {
     ntmp = nin;
   }
   if (nrrdUnaryOpRand == op) {
-    airSrand48(airTime());
+    airSrand48((int)airTime());
   }
   if (nrrdArithUnaryOp(nout, op, ntmp)) {
     airMopAdd(mop, err = biffGetDone(NRRD), airFree, airMopAlways);
